fix(#254): clean up jsonld function and response handling for dataset upload
