+++ resfulservice/src/routes/curation.js
@@ -19,7 +19,4 @@ router.route('/admin')
 router.route('rehydrate')
   .patch(isAuth, curationController.curationRehydration);
 
-router.route('/getdoi/:doi([^/]*)')
-  .get(curationController.getDoiInfo);
-
 module.exports = router;
